+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import HomeCategory from '@/components/HomeCategory';
+import HomeCategory from '@/components/HomePage/HomeCategory/HomeCategory';
 
 export default function HomePage() {
   return (
